Rework console and fmt module
This commit makes the fmt functions able to generate I/O from
generic sources via callbacks. In addition, the console module
has been reworked so that it can work in batches instead of single
characters. The printf function now uses a local buffer instead of
a global one as well.