Implement memory maps for userspace tasks
This changeset extends the functionality of the mapping modules by
allowing the creation and manipulation of non-kernel (i.e: userspace)
memory maps (both physical and virtual). In both cases, the new maps
are created from a template that includes the kernel mappings (which
are inaccessible from userspace, however), but at different boundaries.
Afterwards, they may be manipulated in arbitrary ways.