Implement a DWARF unwinder
This patchset makes use of an architecture agnostic DWARF2 unwinder
to implement stack tracing and unwinding, replacing the 'strace'
module. In addition to being generic, this implementation is much
faster and consumes less stack for fixups when no context needs to be
restored (the more likely path to be taken), while placing a bigger
burden when something goes amiss.