Simplify page fault data retrieval
This commit switches the strategy for retrieving the pages when
a page fault occurs. Instead of using percpu static areas, we
divide objects in 2 groups: Those that use callbacks and those
with external pagers. The former are handled by mapping one page
at a time and calling the callback on that data. The latter are
yet to be implemented.