Implement copy-on-write pages
This changeset implements forking of a VM map, and also COW pages.
In order to do this, the fault routines incorporate the concept
of a 'final' object and offset - i.e: The offset/object pair where
a page will be looked up and inserted. This pair may be different
from the place where the data will be fetched.