Redesign alert delivery in capabilities
Previously, a flow was able to receive 3 types of messages:
synchronous streams, alerts and interrupts. With the upcoming
inclusion of thread and task death delivery, it makes sense to
unify interrupts and alerts. As such, interrupts are no longer
a special kind of message, but are rather delivered as alerts,
only with very high priority.