Use compiler builtins for fixups
The previous implementation use DWARF for stack unwinding, which
is portable but can be sometimes unreliable, and although it's
used only in the slow path, it can be very inefficient. Instead,
this patchset uses the __builtin_{setjmp,longjmp} compiler intrinsics
to implement fixups, which are much more efficient and just as
portable. We'll need to monitor runtime stack usage on architectures
with a large register set (i.e: ARM64) in the future, though.