Re-instate the DWARF unwinder for fixups
After examining the code produced with __builtin_{setjmp,longjmp},
it turns out that the compiler is overly pessimistic, making every
variable access relative to the frame pointer and also saving all
stable registers in the prologue. Since that affects the fast-path,
this commit reverts back to the DWARF unwinder for fixups, which
is much slower in the exceptional path, but doesn't impose any
inefficiencies in the fast one.

In addition, this commit introduces numerous fixes in functions
related to memory management.