Make the unwind mechanism more robust
This patchset adds the frame pointer to the fixup structure. This
is done so that stack frames can be identified more precisely,
instead of playing games with the stack register. New architectures
may not use a frame pointer, and so the code may need to be adjusted
for those newer platforms.