Add support for adjustable priority queues
This PR adds a new data structure: A priority queue that can be
dynamically adjusted so that waiters aren't starved as higher
priority entities are added to it. This can be done in O(1) time,
so the structure is greatly scalable. In addition, it's more
space efficient than the currently used priority lists.