Fix copy-on-write (COW) pages
This patchset fixes the implementation of COW pages. Previously,
we inserted the same page in 2 different objects, which is illegal.
Instead, we now introduce the concept of 'foreign pages': frames that
are present in the physical map, but not in the backing VM object.

In addition, this patchset fixes IPC PTE's so that the previous
physical mappings are saved upon context switch. This makes it so
that these PTEs can be used without disabling interrupts.