Use weak references for channels
In order for servers to be notify when a channel is closed, the
reference count needs to drop to zero. However, the server itself
needs to have the channel open (To get its dirty pages, for example).
As such, this patchset makes it so that servers can make use of
weak references instead, so that their channels aren't accounted
for in the number of references.